Update FinalRectangle.py
Accounted for floating point numbers.

+++ FinalRectangle.py
@@ -1,8 +1,8 @@
 # Problem 5
 class Rectangle:
     def __init__(self, length, width):
-        self.l = length
-        self.w = width
+        self.l = float(length)
+        self.w = float(width)
 
     def perim(self):
         """This function will calculate the perimeter of the Rectangle using the user given values."""
@@ -20,8 +20,8 @@ class Rectangle:
 def rectangle_create():
     """This function creates prompts the user for the length and width of the Rectangle.
         It then performs the necessary calculations and prints the results out to the screen."""
-    userX = int(input("Enter the length of the rectangle:"))
-    userY = int(input("Enter the width of the rectangle:"))
+    userX = float(input("Enter the length of the rectangle:"))
+    userY = float(input("Enter the width of the rectangle:"))
     rect = Rectangle(userX, userY)
     print("Rectangle's Area: " + str(rect.area()))
     print("Rectangle's Perimeter: " + str(rect.perim()))
